Move Int64Tests
