+++ Tests/Batch4/Batch4.csproj
@@ -73,7 +73,6 @@
     <Compile Include="SimpleTypes\TupleTests.cs" />
     <Compile Include="SimpleTypes\SingleTests.cs" />
     <Compile Include="SimpleTypes\SByteTests.cs" />
-    <Compile Include="SimpleTypes\Int64Tests.cs" />
     <Compile Include="SimpleTypes\Int32Tests.cs" />
     <Compile Include="SimpleTypes\Int16Tests.cs" />
     <Compile Include="DelegateTests.cs" />
